Initial commit.
## Rationale

In the initial commit of 1f421, the goal was stated to build Swift
for OpenBSD using a container-based solution, and that repository
contains a reproducible mechanism, namely cloud-init, for obtaining
an operating system container image and being able to orchestrate it
without manual input.

To keep that repository focused, we put all the Swift toolchain build
specifics here, and provides us a place to store Swift toolchain build
artefacts with GitHub Releases and Packages, so we can use them for
bootstrapping purposes.

Using an OpenBSD native build is especially necessary, since there are
platform-specific features in the toolchain and at runtime that need
to be covered. OpenBSD makes customizations to the LLVM toolchain that
may not be immediately upstreamed that can affect binaries at runtime
and so this means cross-compilation is not helpful. Native builds
are the only way to ensure that the toolchain builds and executes.

## Implementation

### Running applications in the virtual machine

cloud-init provides multiple mechanisms to initialize the instance;
for our purposes, the one we are interested in allows us to attach
a disk image to the instance with a special label, `CIDATA`, and
cloud-init detects if such an image is attached to the instance,
reading two YAML files on the image, `meta-data` and `user-data`,
and using the instructions described in those files to configure the
instance. The disk image must either be formatted as FAT or ISO9660.

The necessity to also pass data from the container to the virtual
machine may also present itself. Unfortunately, OpenBSD does not
yet support newer virtio functionality like host to guest 9p mounts
or virtual sockets, so we need to use disk images. We could create
multiple disks: one for the `CIDATA` and one for application data. For
this case, the data is not very large -- we are just bundling a
diff file as part of the application -- but other cases may exist
in the future where we want to transfer much more. For this reason,
ISO9660 is the more attractive solution, since it can handle more
data and longer file names, and the cloud-init yaml files can share
the same disk as the application data, with the small caveat that
it must be read-only. A scratch disk can be created for read-write
temporary storage.

Output data must be read by the Linux container. While OpenBSD and
Linux both support e2fsprogs, which allows for minting ext2 volumes
without mounting, extracting files from an ext2 image without mounting
is more difficult, especially since containers cannot easily mount
disks. Instead, we exploit tar archives: when the VM wants to transmit
data back to the host, it writes the tape archive file directly to
the virtualized disk drive. The host converts the disk drive image
from qcow2 back to raw format and reads the disk image as an ordinary
tape archive. Some care may be required to specify the correct disk
image size to ensure proper format blocking.

We thus need to ensure that we know which disk inside the VM
corresponds to which disk outside the VM, so disks are specified in
the same predictable order: OS image, "scratch" disk, "tape" disk,
then CIDATA.

The steps we want the VM to take need to occur automatically when
the operating system boots. For OpenBSD, rc.firsttime(8) is already
reserved to start cloud-init, but we also have rc.local(8) available
for site-specific commands. This script runs during boot as root, which
leads to some quirks, but this makes the cloud-init configuration
simple: our commands need only be specified as an ordinary script
in rc.local as part of the `user-data` cloud-init configuration:
one that performs our necessary tasks, writes any necessary output
to tape, and then powers off the VM.

Since qemu will still exit successfully if any commands running in the
VM fail, we write the exit code from within the VM to tape and exit
the container explicitly with that exit code, to properly communicate
success or failure as needed.

### The toolchain build application

Perhaps somewhat unconventionally, we tie the actual toolchain
build process to the container build process, rather than building a
container that invokes the build process when the container is run. The
diff file mentioned contains a patch for the new port to be applied to
the ports tree, downloaded inside the virtual machine. The port is then
fully built, the resulting binary packages are output back to tape, and
then stored in a clean scratch container back in the Linux container.

## Caveats

As discussed, since the Swift toolchain build can take several hours
to complete, even on performant machines, the runner should use KVM
and configure the necessary qemu parameters by build argument to match
the build host. Since the build happens inside the virtual machine,
this means builds cannot be easily resumed if interrupted.

narrative: true