Add a GitHub workflow.
## Rationale

Ideally, GitHub Actions could be used to set up a CI pipeline. We
would take the operating system image that we created elsewhere and
then use it to build the port atop that image. Doing this repeatedly
on snapshots would help detect new build problems and changes that
occasionally affect the toolchain build from time to time -- the
last real example of this was OpenBSD making the FILE objects opaque,
which affected the imported Swift-side type.

GitHub Actions would potentially allow binary build artefacts to
be generated in a way such that the provenance of the artefacts
can be better assured. It is much more difficult to interfere
with the process of the build if it runs in a remote cloud and the
artefacts are pushed to a remote container registry where the build
user cannot directly meddle with the generated artefacts while they
are generated. Of course, this shifts the atack surface elsewhere
and the trust and threat scenarios are different, but regardless,
we could be relatively confident in the generation of the artefacts
generated through such a workflow.

## Caveats

GitHub offers hosted runners that execute the container build as
configured with the free tier being 4 cores and 16GB of RAM. However,
with the runner dimensioned this way, the build takes more than six
hours to complete. This means the container will never build. GitHub
offers larger runners behind a paid tier, but there are reliability
issues immediately apparent with this service.

For toolchain bootstrapping, it is more ideal for the port to refer
to a GitHub Package than a container image, since the container layer
cannot be downloaded trivially with ftp(1); it requires an explicit
header to be passed along. The port could potentially use curl, but
that requires somewhat unusual configuration for the port. Other
OpenBSD ports use GitHub packages in this same way. This requires
binary tarballs to be uploaded manually.

If we are going to use Packages in this way, then we have to have this
manual step, and the artefact generation pipeline from code to binary
would be broken anyway, so having a GitHub container build workflow is
really only useful for automated port build testing on newer operating
system images. If larger runners don't seem to schedule work well,
then why bother? We will just create the build artefacts locally and
upload them as Packages.

This exercise is however useful as a reference, so this commit will
remain. There may be some tricks we can do to persist build artefacts
so we can potentially resume the build before the first six hours
are up, but that's not important right now. Or perhaps larger runners
would work properly in the future.

narrative: true